+++ render/software/src/portals.rs
@@ -5,23 +5,23 @@ pub struct PortalClip {
     /// Clip values are the solid pixel bounding the range.
     ///  floorclip starts out SCREENHEIGHT
     ///  ceilingclip starts out -1
-    pub floorclip: Vec<f32>,
-    pub ceilingclip: Vec<f32>,
+    pub floorclip: Vec<i32>,
+    pub ceilingclip: Vec<i32>,
     screen_height: usize,
 }
 
 impl PortalClip {
     pub fn new(screen_width: usize, screen_height: usize) -> Self {
         PortalClip {
-            floorclip: vec![0.0; screen_width + 1],
-            ceilingclip: vec![0.0; screen_width + 1],
+            floorclip: vec![0; screen_width + 1],
+            ceilingclip: vec![0; screen_width + 1],
             screen_height,
         }
     }
 
     pub(super) fn clear(&mut self) {
-        self.floorclip.fill(self.screen_height as f32);
-        self.ceilingclip.fill(-1.0);
+        self.floorclip.fill(self.screen_height as i32);
+        self.ceilingclip.fill(-1);
     }
 }
